Android: Updated changes for [TIMOB-24537] based on code review feedback.
- Removed undocumented/deprecated Ti.Gesture APIs isPortrait() and isLandscape().
- Removed Java "TiOrientationHelper" fclass. Replaced by new "TiDeviceOrientation" class.
- Removed deprecated TiUIHelper orientation constants. Replaced with TiDeviceOrientation class constants.
- Changed UNKNOWN orientation integer ID from 7 to 0.
- Modified "Window.orientationModes" setter to validate and log warnings for invalid identifiers.
